renamed visualize.py
Renamed visualize.py to show.py
